localization-management-api/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
supabase
pytest
pytest-asyncio